requirements.txt
light-embed
fastapi
uvicorn
pydantic
python-dotenv
groq
scipy
pinecone
